+++ include/Gaffer/Animation.h
@@ -54,7 +54,7 @@ class GAFFER_API Animation : public ComputeNode
 
 	public :
 
-		Animation( const std::string &name=defaultName<Animation>() );
+		explicit Animation( const std::string &name=defaultName<Animation>() );
 		~Animation() override;
 
 		GAFFER_NODE_DECLARE_TYPE( Gaffer::Animation, AnimationTypeId, ComputeNode );
@@ -62,17 +62,142 @@ class GAFFER_API Animation : public ComputeNode
 		/// Defines the method used to interpolate between a key and the next one.
 		enum class Interpolation
 		{
-			Constant,
+			Constant = 0,
 			ConstantNext,
 			Linear,
-			/// \todo Add Smooth, implemented as
-			/// bezier curves using V2f in and out
-			/// tangents on each key.
+			Cubic,
+			Bezier
 		};
 
+		/// Defines direction relative to a key.
+		enum class Direction
+		{
+			In = 0,
+			Out
+		};
+
+		/// Defines whether slope and scale are tied.
+		enum class TieMode
+		{
+			Manual = 0,
+			Slope,
+			Scale
+		};
+
+		/// Get the default interpolation mode.
 		static Interpolation defaultInterpolation();
 
+		/// Get the default tie mode.
+		static TieMode defaultTieMode();
+
+		/// Get the opposite direction to the specified direction
+		static Direction opposite( Direction direction );
+
+		/// Get the default slope
+		static double defaultSlope();
+
+		/// Get the default scale
+		static double defaultScale();
+
+		class Key;
 		class CurvePlug;
+		class Interpolator;
+		IE_CORE_DECLAREPTR( Interpolator )
+
+		// Defines a tangent
+		class Tangent : private boost::noncopyable
+		{
+			public:
+
+				~Tangent();
+
+				/// Get parent key.
+				Key& key();
+				/// Get parent key. (const access)
+				const Key& key() const;
+
+				/// Get the direction of the tangent
+				Direction direction() const;
+
+				/// Get tangent's slope.
+				/// The slope is in range [-inf,+inf]
+				/// If slopeIsConstrained() returns true this function will return the constrained slope.
+				double getSlope() const;
+				/// \undoable
+				/// Set tangent's slope.
+				/// The slope is in range [-inf,+inf]
+				/// If the tangent's key has tie mode set to either Slope or Scale the opposite tangents slope will be set to the same value.
+				/// If slopeIsConstrained() returns true this function will have no effect.
+				void setSlope( double slope );
+				/// \undoable
+				/// Set tangent's position from specified position whilst maintaining the current scale.
+				/// If relative is true the position is relative to the parent key's position.
+				/// If the tangent's key has tie mode set to either Slope or Scale the opposite tangents slope will be set to the same value.
+				/// If slopeIsConstrained() returns true this function will have no effect.
+				/// Slope cannot be set from a position if there is no adjacent key in the direction of the tangent.
+				void setSlopeFromPosition( const Imath::V2d& position, bool relative = false );
+
+				/// Get tangent's scale.
+				/// The scale is multiplied by the span width to derive the tangent's length.
+				/// If scaleIsConstrained() returns true this function will return the constrained scale.
+				double getScale() const;
+				/// \undoable
+				/// Set tangent's scale.
+				/// The scale is multiplied by the span width to derive the tangent's length.
+				/// If the tangent's key has tie mode set to Scale the opposite tangent's scale will be kept proportional.
+				/// If scaleIsConstrained() returns true this function will have no effect.
+				void setScale( double scale );
+				/// \undoable
+				/// Set tangent's scale from the specified position whilst maintaining the current slope.
+				/// If relative is true the position is relative to the parent key's position.
+				/// If the tangent's key has tie mode set to Scale the opposite tangent's scale will be kept proportional.
+				/// If scaleIsConstrained() returns true this function will have no effect.
+				/// Scale cannot be set from a position if there is no adjacent key in the direction of the tangent.
+				void setScaleFromPosition( const Imath::V2d& position, bool relative = false );
+
+				/// \undoable
+				/// Set tangent's slope and scale, constrained slope and/or scale will be maintained.
+				void setSlopeAndScale( double slope, double scale );
+
+				/// Is slope currently constrained by interpolation mode.
+				bool slopeIsConstrained() const;
+				/// Is scale currently constrained by interpolation mode.
+				bool scaleIsConstrained() const;
+
+				/// Get tangent's position.
+				/// If relative is true the position is relative to the parent key's position.
+				/// The position will be derived from the current (possibly constrained) slope and scale.
+				/// The position will be the same as the parent key when there is no adjacent key in the direction of the tangent.
+				Imath::V2d getPosition( bool relative = false ) const;
+				/// \undoable
+				/// Set tangent's position.
+				/// If relative is true the position is relative to the parent key's position.
+				/// The position will be used to derive a new slope and scale for the tangent, constrained slope and/or scale will be maintained.
+				/// The position cannot be set if there is no adjacent key in the direction of the tangent.
+				void setPosition( const Imath::V2d& position, bool relative = false );
+
+			private:
+
+				friend class CurvePlug;
+				friend class Key;
+
+				Tangent( Key&, Direction, double, double );
+				/// \undoable
+				void setSlope( double, bool );
+				/// \undoable
+				void setScale( double, bool );
+				/// \undoable
+				void setSlopeAndScale( double, double, bool );
+				void update();
+				void positionToRelative( Imath::V2d&, bool ) const;
+
+				Key* m_key;
+				Direction m_direction;
+				double m_dt;
+				double m_dv;
+				double m_slope;
+				double m_scale;
+		};
 
 		/// Defines a single keyframe.
 		class Key : public IECore::RunTimeTyped
@@ -80,10 +205,37 @@ class GAFFER_API Animation : public ComputeNode
 
 			public :
 
-				explicit Key( float time = 0.0f, float value = 0.0f, Interpolation interpolation = Animation::defaultInterpolation() );
+				explicit Key( float time = 0.0f, float value = 0.0f, Interpolation interpolation = Animation::defaultInterpolation(),
+					double inSlope = Animation::defaultSlope(), double inScale = Animation::defaultScale(),
+					double outSlope = Animation::defaultSlope(), double outScale = Animation::defaultScale(),
+					TieMode tieMode = Animation::defaultTieMode() );
 				~Key() override;
 
 				IE_CORE_DECLARERUNTIMETYPEDEXTENSION( Gaffer::Animation::Key, AnimationKeyTypeId, IECore::RunTimeTyped )
+
+				// Get in tangent
+				Tangent& tangentIn();
+				// Get in tangent (const access)
+				const Tangent& tangentIn() const;
+				// Get out tangent
+				Tangent& tangentOut();
+				// Get out tangent (const access)
+				const Tangent& tangentOut() const;
+				// Get tangent in specified direction
+				Tangent& tangent( Direction direction );
+				// Get tangent in specified direction (const access)
+				const Tangent& tangent( Direction direction ) const;
+
+				/// Get current tie mode of key.
+				TieMode getTieMode() const;
+				/// Set tie mode of key. If tie mode is Slope or Scale the slope of the in and
+				/// out tangents with be made equal. If only one tangent's slope is constrained
+				/// or protrudes beyond the start or end of the parent curve, the opposite tangent's
+				/// slope will be preserved, otherwise the slopes are averaged. If tie mode is Scale
+				/// the ratio between the in and out tangent's scales is captured and changes to
+				/// either tangent's scale preserve the proportionality of the opposite tangent's scale.
+				/// \undoable
+				void setTieMode( TieMode mode );
 
 				/// Get current time of key.
 				float getTime() const;
@@ -109,12 +261,20 @@ class GAFFER_API Animation : public ComputeNode
 				/// Is the key currently active? The key is considered inactive whilst unparented.
 				bool isActive() const;
 
+				/// Get parent curve.
 				CurvePlug *parent();
+				/// Get parent curve (const access).
 				const CurvePlug *parent() const;
 
 			private :
 
 				friend class CurvePlug;
+				friend class Tangent;
+
+				Key *nextKey();
+				Key *prevKey();
+				const Key *nextKey() const;
+				const Key *prevKey() const;
 
 				void throwIfStateNotAsExpected( const CurvePlug*, bool, float ) const;
 
@@ -132,11 +292,17 @@ class GAFFER_API Animation : public ComputeNode
 					void operator()( Key* ) const;
 				};
 
+				static Tangent Key::* const m_tangents[ 2 ];
+
 				Hook m_hook;
 				CurvePlug *m_parent;
+				Tangent m_tangentIn;
+				Tangent m_tangentOut;
 				float m_time;
 				float m_value;
-				Interpolation m_interpolation;
+				ConstInterpolatorPtr m_interpolator;
+				double m_tieScaleRatio;
+				TieMode m_tieMode;
 				bool m_active;
 
 		};
@@ -145,6 +311,9 @@ class GAFFER_API Animation : public ComputeNode
 
 		class KeyIterator;
 		class ConstKeyIterator;
+
+		class Extrapolator;
+		IE_CORE_DECLAREPTR( Extrapolator )
 
 		/// Defines a curve as a collection of keyframes and methods
 		/// for editing them. Provides methods for evaluating the
@@ -156,7 +325,7 @@ class GAFFER_API Animation : public ComputeNode
 
 				GAFFER_PLUG_DECLARE_TYPE( Gaffer::Animation::CurvePlug, AnimationCurvePlugTypeId, Gaffer::ValuePlug );
 
-				CurvePlug( const std::string &name = defaultName<CurvePlug>(), Direction direction = Plug::In, unsigned flags = Plug::Default );
+				explicit CurvePlug( const std::string &name = defaultName<CurvePlug>(), Plug::Direction direction = Plug::In, unsigned flags = Plug::Default );
 				~CurvePlug() override;
 
 				typedef boost::signal< void ( CurvePlug*, Key* ), Gaffer::CatchingSignalCombiner< void > > CurvePlugKeySignal;
@@ -166,6 +335,7 @@ class GAFFER_API Animation : public ComputeNode
 				CurvePlugKeySignal& keyTimeChangedSignal();
 				CurvePlugKeySignal& keyValueChangedSignal();
 				CurvePlugKeySignal& keyInterpolationChangedSignal();
+				CurvePlugKeySignal& keyTieModeChangedSignal();
 
 				/// Adds specified key to curve, if key is parented to another curve or already parented
 				/// to the curve and inactive it is removed from its parent curve. If the key has already
@@ -244,6 +414,7 @@ class GAFFER_API Animation : public ComputeNode
 				/// iterator to end of range of active keys. (const access)
 				ConstKeyIterator end() const;
 
+				/// Evaluate the curve at the specified time
 				float evaluate( float time ) const;
 
 				/// Output plug for evaluating the curve
@@ -255,8 +426,11 @@ class GAFFER_API Animation : public ComputeNode
 			private :
 
 				friend class Key;
+				friend class Tangent;
 				friend KeyIterator;
 				friend ConstKeyIterator;
+
+				typedef boost::signal< void ( CurvePlug*, Animation::Direction ), Gaffer::CatchingSignalCombiner< void > > CurvePlugDirectionSignal;
 
 				Key *firstKey();
 				Key *finalKey();
@@ -268,7 +442,7 @@ class GAFFER_API Animation : public ComputeNode
 				struct TimeKey
 				{
 					typedef float type;
-					type operator()( const Animation::Key& key ) const; // NOTE : must NEVER throw
+					type operator()( const Animation::Key& ) const; // NOTE : must NEVER throw
 				};
 
 				typedef boost::intrusive::member_hook< Key, Key::Hook, &Key::m_hook > KeyHook;
@@ -277,6 +451,8 @@ class GAFFER_API Animation : public ComputeNode
 				typedef boost::intrusive::avl_set< Key, KeyHook, KeyOfValue > Keys;
 				typedef boost::intrusive::avl_multiset< Key, KeyHook, KeyOfValue > InactiveKeys;
 
+				static ConstExtrapolatorPtr CurvePlug::* const m_extrapolators[ 2 ];
+
 				Keys m_keys;
 				InactiveKeys m_inactiveKeys;
 				CurvePlugKeySignal m_keyAddedSignal;
@@ -284,10 +460,16 @@ class GAFFER_API Animation : public ComputeNode
 				CurvePlugKeySignal m_keyTimeChangedSignal;
 				CurvePlugKeySignal m_keyValueChangedSignal;
 				CurvePlugKeySignal m_keyInterpolationChangedSignal;
+				CurvePlugKeySignal m_keyTieModeChangedSignal;
+				CurvePlugDirectionSignal m_extrapolationChangedSignal;
+				ConstExtrapolatorPtr m_extrapolatorIn;
+				ConstExtrapolatorPtr m_extrapolatorOut;
 		};
 
 		/// convert enums to strings
 		static const char* toString( Interpolation interpolation );
+		static const char* toString( Direction direction );
+		static const char* toString( TieMode mode );
 
 		IE_CORE_DECLAREPTR( CurvePlug );