+++ include/GafferUI/AnimationGadget.h
@@ -101,6 +101,8 @@ class GAFFERUI_API AnimationGadget : public Gadget
 		void removeInactiveKeyframes();
 		/// \undoable
 		void moveKeyframes( const Imath::V2f currentDragOffset );
+		/// \undoable
+		void moveTangent( const Imath::V2f currentDragOffset );
 
 		void frame();
 
@@ -121,6 +123,8 @@ class GAFFERUI_API AnimationGadget : public Gadget
 		// Find elements at certain positions
 		Gaffer::Animation::ConstKeyPtr keyAt( const IECore::LineSegment3f &position ) const;
 		Gaffer::Animation::KeyPtr keyAt( const IECore::LineSegment3f &position );
+		std::pair<Gaffer::Animation::ConstKeyPtr, Gaffer::Animation::Direction> tangentAt( const IECore::LineSegment3f &position ) const;
+		std::pair<Gaffer::Animation::KeyPtr, Gaffer::Animation::Direction> tangentAt( const IECore::LineSegment3f &position );
 		Gaffer::Animation::ConstCurvePlugPtr curveAt( const IECore::LineSegment3f &position ) const;
 		Gaffer::Animation::CurvePlugPtr curveAt( const IECore::LineSegment3f &position );
 		bool frameIndicatorUnderMouse( const IECore::LineSegment3f &position ) const;
@@ -148,6 +152,10 @@ class GAFFERUI_API AnimationGadget : public Gadget
 
 		std::map< const Gaffer::Animation::Key*, std::pair< float, float > > m_originalKeyValues;
 
+		Gaffer::Animation::KeyPtr m_dragTangentKey;
+		Gaffer::Animation::Direction m_dragTangentDirection;
+		double m_dragTangentOriginalScale;
+
 		Imath::V2f m_dragStartPosition;
 		Imath::V2f m_lastDragPosition;
 
@@ -156,7 +164,8 @@ class GAFFERUI_API AnimationGadget : public Gadget
 			None,
 			Selecting,
 			Moving,
-			MoveFrame
+			MoveFrame,
+			MoveTangent
 		};
 
 		DragMode m_dragMode;
@@ -174,6 +183,8 @@ class GAFFERUI_API AnimationGadget : public Gadget
 		Gaffer::Animation::KeyPtr m_snappingClosestKey;
 		Gaffer::Animation::KeyPtr m_highlightedKey;
 		Gaffer::Animation::CurvePlugPtr m_highlightedCurve;
+		Gaffer::Animation::KeyPtr m_highlightedTangentKey;
+		Gaffer::Animation::Direction m_highlightedTangentDirection;
 
 		int m_mergeGroupId;